Add FavoriteDisplay, Update Complex grid, Delete Stars and Adjusted Stars
