+++ yelp/src/components/FavoritesDisplay.js
@@ -1,42 +1,25 @@
-import React, { useState } from 'react';
+import React from 'react'
 import {
-  Card, CardText, CardBody, CardLink,
-  CardTitle, CardSubtitle, UncontrolledCollapse, Button, TabContent, TabPane,
-   Nav, NavItem, NavLink, Row, Col
-} from 'reactstrap';
-import classnames from 'classnames';
-
-import NAVBAR from './Mode'
-import PropTypes from 'prop-types';
-import { makeStyles } from '@material-ui/core/styles';
-import Rating from '@material-ui/lab/Rating';
-import Tooltip from '@material-ui/core/Tooltip';
-import Typography from '@material-ui/core/Typography';
-import Box from '@material-ui/core/Box';
-
-
-
-const useStyles = makeStyles({
-  rating1: {
-    width: 200,
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    flexDirection: "column",
-    margin: "0, auto",
-    padding: "1rem",
-    fontSize: "2rem",
-  },
-});
+    Card, CardText, CardBody, CardLink,
+    CardTitle, CardSubtitle, UncontrolledCollapse, Button, TabContent, TabPane,
+     Nav, NavItem, NavLink, Row, Col
+  } from 'reactstrap';
+  import classnames from 'classnames';
+  
+  import NAVBAR from './Mode'
+  import PropTypes from 'prop-types';
+  import { makeStyles } from '@material-ui/core/styles';
+  import Rating from '@material-ui/lab/Rating';
+  import Tooltip from '@material-ui/core/Tooltip';
+  import Typography from '@material-ui/core/Typography';
+  import Box from '@material-ui/core/Box';
 
 
+const FavoriteDisplay = (props)  => {
+    console.log(`FavoriteDisplay`,props)
 
-const ComplexGrid = (props) => {
-    // const [activeTab, setActiveTab] = useState('1');
 
-    console.log(`props.fact.user_review`,props.fact)
- 
-  return (
+    return(
     <>
     <div>
       <NAVBAR />
@@ -57,7 +40,7 @@ const ComplexGrid = (props) => {
         <div className="stars">
           <Rating
             name="hover-side"
-            value={props.fact.user_review}
+            value={props.fact.original_yelp_user_rating}
             readOnly = {true}
           />
         </div>
@@ -77,7 +60,7 @@ const ComplexGrid = (props) => {
         <div className="stars">
           <Rating
             name="hover-side"
-            value={props.fact.yelp_rating}
+            value={props.fact.yelp_store_rating}
             readOnly = {true}
           />
         </div>
@@ -88,4 +71,4 @@ const ComplexGrid = (props) => {
   );
 };
 
-export default ComplexGrid;
+export default FavoriteDisplay
