+++ yelp/src/components/Review.jsx
@@ -2,7 +2,8 @@ import React, { useEffect } from 'react';
 import { connect } from 'react-redux';
 import { fetchRest } from '../actions';
 import YDetail from './YDetail'
-
+import HoverRating from './Stars'
+import AdjustedRating from './AdjustedStars'
 
 const Review = props => {
     
@@ -19,8 +20,13 @@ const Review = props => {
     <div>
       {props.error && <p>{props.error}</p>}
       {props.diner.map(fact => (
+         <React.Fragment>
         <YDetail key={fact.business_id} fact={fact} />
-      ))}
+        {/* <HoverRating key={fact.business_id} fact={fact} />
+        <AdjustedRating key={fact.business_id} fact={fact} /> */}
+        </React.Fragment>
+      ))
+      }
     </div>
   );
 };
