+++ yelp/src/components/SimpleGrid.jsx
@@ -3,7 +3,29 @@ import {
   Card, CardText, CardBody, CardLink,
   CardTitle, CardSubtitle
 } from 'reactstrap';
-import AdjustedRating from './AdjustedStars';
+
+import NAVBAR from './Mode'
+import PropTypes from 'prop-types';
+import { makeStyles } from '@material-ui/core/styles';
+import Rating from '@material-ui/lab/Rating';
+import Tooltip from '@material-ui/core/Tooltip';
+import Typography from '@material-ui/core/Typography';
+import Box from '@material-ui/core/Box';
+
+const useStyles = makeStyles({
+  rating1: {
+    width: 200,
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+    flexDirection: "column",
+    margin: "0, auto",
+    padding: "1rem",
+    fontSize: "2rem",
+  },
+});
+
+
 
 const SimpleGrid = props => {
   console.log(props)
@@ -21,6 +43,18 @@ const SimpleGrid = props => {
           {/* <CardLink className="stars" href="#"><AdjustedRating /></CardLink> */}
         </CardBody>
       </Card>
+      <div id="myDIV">
+      <Box component="fieldset" mb={3} borderColor="transparent">
+        <Typography component="legend"><h1>Adjusted Rating:</h1></Typography>
+        <div className="stars">
+          <Rating
+            name="hover-side"
+            value={props.fact.yelp_rating}
+            readOnly = {true}
+          />
+        </div>
+      </Box>
+      </div>
     </div>
   );
 };
