+++ yelp/src/components/YDetail.jsx
@@ -3,7 +3,8 @@ import React from 'react'
 import ComplexGrid from "./ComplexGrid";
 //import SimpleGrid from "./SimpleGrid";
 import { Link } from "react-router-dom";
-import AdjustedRating from './AdjustedStars';
+
+
 
 //create an individual link on each item,
 // uses /movies/fact._id -> extract id, and add to movies to create new url
@@ -15,7 +16,6 @@ function YDetail({ fact }) {
       <Link to={`/reviews/${fact.id}`}>
         {/* <ProfileCard fact={fact} /> */}
         <ComplexGrid fact={fact} />
-        <AdjustedRating />
       </Link>
       </>
     );
